Cleaner sound conversion based on new VGAudioCli
